refactor: extract observation artifact helpers
